CogVM source as per VMMaker.oscog-eem.1602
Spur: Missing classArray export.

Don't include sqMemoryAccess.h include explicitly since it's pulled in by sq.h anyway.

Fix external declarations for several functions used in the interpreter that were
previously undeclared.

Fix include paths for the (moved) gdb ARM support files.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3542 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 40b18ac68cd5ad7f64b08cfddc362c48728f4d08

+++ platforms/unix/vm/aio.c
@@ -204,7 +204,7 @@ do if ((bool) && !(++tickCount % TICKS_PER_CHAR)) {		\
 	if (!*ticker++) ticker= ticks;			\
 } while (0)
 
-int aioPoll(int microSeconds)
+long aioPoll(long microSeconds)
 {
   int	 fd;
   fd_set rd, wr, ex;
@@ -273,7 +273,7 @@ int aioPoll(int microSeconds)
 /* sleep for microSeconds or until i/o becomes possible, avoiding
    sleeping in select() if timeout too small */
 
-int aioSleepForUsecs(int microSeconds)
+long aioSleepForUsecs(long microSeconds)
 {
 #if defined(HAVE_NANOSLEEP)
   if (microSeconds < (1000000/60))	/* < 1 timeslice? */
